backend_django/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.5.0
python-decouple==3.8
django-cors-headers==4.3.0

